+++ src/main/java/mcjty/lib/varia/IEnergyItem.java
@@ -3,11 +3,11 @@ package mcjty.lib.varia;
 import net.minecraft.item.ItemStack;
 
 public interface IEnergyItem {
-    int receiveEnergy(ItemStack container, int maxReceive, boolean simulate);
+    long receiveEnergy(ItemStack container, long maxReceive, boolean simulate);
 
-    int extractEnergy(ItemStack container, int maxExtract, boolean simulate);
+    long extractEnergy(ItemStack container, long maxExtract, boolean simulate);
 
-    int getEnergyStored(ItemStack container);
+    long getEnergyStored(ItemStack container);
 
-    int getMaxEnergyStored(ItemStack container);
+    long getMaxEnergyStored(ItemStack container);
 }
